refactor: updated metric, data, dto, mapper, and other gas-resistance references in BE;
TODO: FE

+++ src/main/java/at/qe/skeleton/dtos/MeasurementDTO.java
@@ -8,7 +8,7 @@ import java.util.List;
 public record MeasurementDTO(
         Long id,
         LocalDateTime timestamp,
-        Long measurement,
+        Float measurement,
         Metric metric,
         Long roomId,
         Long sensorStationId,
